requirements.txt
flask==1.0.2
coverage==4.5.3
mysql-connector==2.2.9
pymongo==3.7.2
pytest==4.4.0